+++ src/main/java/org/bytedeco/javacpp/annotation/ByRef.java
@@ -20,6 +20,6 @@ import org.bytedeco.javacpp.tools.Generator;
 @Documented @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.PARAMETER})
 public @interface ByRef {
-    /** A C++ code snippet to use when null is passed. By default {@link NullPointerException} is thrown. */
+    /** A C++ expression to use when null is passed. By default {@link NullPointerException} is thrown. */
     String nullValue() default "";
 }